requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
gunicorn==22.0.0
python-multipart==0.0.9
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.4.0
pydantic==2.8.2
pydantic-settings==2.3.4
